Make `commands` optional in run configurations (#1104)
Now, if the user specifies a custom image in
their configuration, they can omit `commands` so
that the image's default `ENTRYPOINT`+`CMD` and
default working directory are used.

We always override the image's entrypoint, so it
is necessary to find out what the default
entrypoint was and pass it to the runner. This is
done by the dstack server by requesting the
image's manifest from its container registry.