[Internal] Add script for checking OCI images (#1408)
This new command in `oci_image_tools.py` checks
that images are published in the OCI Marketplace
and completed the OCI review successfully. This
command is supposed to be used manually. Using it
as part of the images build CI workflow is not
viable, as reviews can take indefinite time.