Approximate on-prem GPU memory size
If the GPU memory size reported by `nvidia-smi` on
an on-prem instance differs slightly from the
memory this GPU model is known to have, `dstack`
will now use the known memory.

Additionally, this commit adjusts the translation
of on-prem GPU names for some edge cases so that
more on-prem GPU names match cloud GPU names and
are recognized during memory approximation.