Mitigate concurrent `dstack attach` issues
1. Don't use SO_REUSEADDR. Although it might be helpful, as it allows to
   bind a socket in TIME_WAIT state (this is the reason it was added in
   the first place), unfortunately it has some other (unwanted) effects:
    * On Linux, it allows to bind to the same address-port pair
      (including wildcard), as long as this option is set before each
      bind().
    * On BSD, it allows to bind to a wildcard address and a specific
      address, or vice versa (but not the same specific address, as on
      Linux).
These other effects increase discrepancy between Linux and BSD (incl.
macOS).

2. Run ssh with ExitOnForwardFailure=yes. With this option, ssh exits
   with error if it cannot bind() to requested ports.

This is not an ideal solution. the way Run.attach() works, it's still
possible to PortsLock.acquire() the same local port due to race condition
(if one client just released the lock, but ssh hasn't yet established
the tunnel, another client can acquire the same port in between, but
with the second fix applied it will eventually fail, as it will not be
able to establish ssh tunnel in 10 attempts.

It would be better to acquire PortsLock as soon as possible (don't wait
for RunStatus.RUNNING), but it requires refactoring, including the
public Python API.

Fixes: https://github.com/dstackai/dstack/issues/1814