Support gateway features in dstack-proxy
This commit adds all gateway features to
dstack-proxy: registry API, config API, stats API,
Nginx as reverse proxy, automatic certificate
issuance. This is in preparation for running
dstack-proxy on gateways, which will be enabled
later in a separate PR. So far, the changes do not
affect the use of dstack in any way.

Most of the code is taken from dstack-gateway and
adapted to the dstack-proxy architecture. Some
notable changes compared to dstack-gateway:
- Services are identified by their project name
  and run name, not run id. This allows the
  gateway and in-server environments to share more
  code. The registry and stats APIs are changed
  accordingly.
- All state is concentrated in the repo -
  singletons like Nginx or StatsCollector don't
  have persistent state.
- Since the Nginx singleton is now stateless,
  Nginx configs are constructed from the service
  state on each change, rather than keeping the
  current config version in memory and applying
  changes to it.
- Unit test coverage.

The commit also includes adjustments to the
dstack-proxy architecture influenced by the
addition of new features:
- Auth checks are moved from BaseProxyRepo to a
  separate dependency, BaseAuthProvider.
- BaseProxyRepo implementations are renamed to
  ServerProxyRepo and GatewayProxyRepo.
- The dependency injection mechanism is adjusted
  to allow gateway-only API routes to depend
  directly on GatewayProxyRepo. This allows to
  define gateway-specific repo methods in
  GatewayProxyRepo only. API routes common to both
  the gateway and in-server environments still
  depend on BaseProxyRepo.

The following is not part of this commit and will
be added later:
- Actually running dstack-proxy on gateways.
- Restoring the application state after restarts.
- Server-side adjustments to accommodate for the
  registry and stats API changes.