Optimize gateway startup and service update time
Avoid running certbot if the certificate exists.
This greatly reduces the gateway startup time when
many services or entrypoints are being
re-registered. Running certbot can take 3-4
seconds, even if the certificate already exists
and there is nothing to do.

Certificate renewal is not an issue as it is done
by certbot's built-in systemd timer.