Fix Fabric Manager in AWS/GCP/Azure/OCI OS images
The `cuda-drivers-535` and
`nvidia-fabricmanager-535` packages used to be
pinned to a specific version. However, this didn't
actually pin the driver version, because
`cuda-drivers-535` brings several NVIDIA packages
and allows various versions of them.

```
Package: cuda-drivers-535
Version: 535.183.01-1
...
Depends: libnvidia-common-535 (>= 535.183.01), libnvidia-compute-535 (>= 535.183.01), ..., nvidia-driver-535 (>= 535.183.01), ...
```

This could lead to a version mismatch when a newer
driver version is installed while Fabric Manager
is pinned to an older version.

This commit solves the issue by removing the pin
so that both the driver and the Fabric Manager
have the same latest `535.*` version.