Try more offers when starting a job
Increase the number of offers tried from 15 to 25.
Also make this configurable.

In particular, this is needed for RunPod Community
Cloud spot offers, because availability detection
is imprecise there, so many offers may turn out to
be unavailable. The increased limit can be seen as
a temporary measure until we improve the
availability detection in RunPod.