Replace sync SDK API with a custom wrapper
The sync API suffers from deadlocks, as stated in
the SDK readme. Deadlocks often happen during
concurrent actions, such as creating or deleting
fleets with many instances, and lead to a
dysfunctional dstack-server. This commit solves
the problem by switching to the async API and
running the async methods in a separate event loop
dedicated to Nebius.