Fix request filtering for service stats (#2678)
Instead of relying on status codes or other
heuristics, have Nginx log a flag for each request
forwarded, or intended to be forwarded, to a
service replica. Use these flags to determine
which requests to include in service stats.

This reliably excludes requests that are
terminated by Nginx and hence don't add load to
replicas. These include HTTP to HTTPS redirects,
unauthenticated requests, and requests that
violate rate or body size limits.