Add `files` property to run configurations
Each item in the `files` property maps a local path (a file or a dir)
to a container path. Each local path is packed into a tar archive and
uploaded to the server similar to the code repo archive/diff.

On the server, each archive is stored in the DB as `FileArchiveModel`
linked to the user. Archive blobs are optionally uploaded to
the storage (again, similar to the code blob).

When the job is submitted to the runner, files (if any) are uploaded
after `/api/submit` but before `/api/upload_code`.

The runner unpacks archives as follows:
* If the path already exists, it's removed
* If any parent dirs of the path are missing, they are created as owned
  by the run user. The owner of the existing dirs is not changed.
* The owner of the path (and all subpaths in the case of the directory)
  is set to the run user. The permissions from the archive (and thus
  from the user's machine) are preserved.

Part-of: https://github.com/dstackai/dstack/issues/2738