Rolling deployments for `files` (#2862)
Support rolling deployments for services when the
`files` property is updated or the contents of
file archives change.

Currently, `dstack` always detects phantom changes
to file archive contents, since file archives are
omitted when getting the run plan. Addressing this
is out of scope for this PR.

This PR does not address compatibility with
previously submitted jobs that are using files,
because the files feature has not been released
yet, so no users have such jobs.