Fix NVIDIA container toolkit bug in all backends
Use cgroupfs as a Docker cgroup driver on all
backends by default to work around an NVIDIA
container toolkit bug. The patch to
`/etc/docker/daemon.json` is automatically applied
in all VM-based backends if
`/etc/docker/daemon.json` exists and `jq` is
installed. This is not the case in Nebius. Lambda,
and CUDO, so they still need custom code to apply
the workaround.