Fix NVIDIA container toolkit bug in all backends (#2877)
Use cgroupfs as a Docker cgroup driver on all
backends by default to work around an NVIDIA
container toolkit bug where the container looses
access to the GPU.

The patch to `/etc/docker/daemon.json` is
automatically applied in all VM-based backends if
`/etc/docker/daemon.json` exists, has the NVIDIA
runtime, does not explicitly set another cgroup
driver, and `jq` is installed. This is not the
case in Nebius. Lambda, and CUDO, so they still
need custom code to apply the workaround - either
installing `jq` or just writing a hardcoded
`/etc/docker/daemon.json` that is known to work on
this backend.