Add/remove SSH instances via in-place update (#2884)
It's now possible to add/remove instances to/from SSH fleets applying
the fleet configuration with adjusted `ssh_config.hosts`.

Currently, the old and the new configurations must be equal except for
the `identity_file` fields (ignored) and the order of hosts in
the `ssh.config` list (but each host must remain the same or
be removed).

Closes: https://github.com/dstackai/dstack/issues/2645