[shim] Don't check image downloaded size (#2903)
Sometimes dockerd emits less "Download complete" messages than
expected, but the image is pulled successfully.
The reason is unclear, but, anyway, this check is redundant since
we also rely on the status message, which is emitted only
in the case of succesfull pull.

In addition, this patch adds/changes the following:

* Write dockerd pull stream (JSON Lines) to {runnerDir}/pull.log —
  useful for debugging (in conjunction with
  DSTACK_SERVER_KEEP_SHIM_TASKS=1)
* Use `errorDetail.message` instead of deprecated `error`
* Move `ctx.Err()` check upper, otherwise it's shadowed by pull errors

Fixes: https://github.com/dstackai/dstack/issues/2503