Fix SSHAttach.detach()
When error logging was added to SSHTunnel.close()[1], it was discovered
that SSHAttach.detach() is called multiple times on detach.
Although this is harmless as long as this method is idempotent (and it is),
this patch ensures that detach() is no-op when called more than one
time, removing ERROR level log noise.

As a side-effect, detaching SSHAttach that reuses SSHTunnel no longer
kills "master" SSHAttach, meaning that one can `dstack attach` multiple
times and detach independently (detaching from the "master" attach still
kicks out all other attaches).

[1]: https://github.com/dstackai/dstack/pull/3296