Add `EventTarget.project_name`
Return the project name for all targets in
`/api/events/list`. Required for displaying event
targets and building entity URLs in the UI.

Implemented by joining event targets with
projects. The performance effect of the join are
tolerable — with a dataset of 1.8 million events,
requests with various filters take about 0.05s
with hot Postgres cache, and up to ~0.75s with
cold cache, which is similar to previous
measurements.