Revert job-related changes
To avoid extra testing and changes, limit the
scope of the PR to instance termination reasons,
as per the PR name.

The introduction of
`JobTerminationReason.NO_BALANCE` can be done in a
separate PR. Its implementation may need to be
updated to also use
`JobTerminationReason.NO_BALANCE` on jobs with
statuses other than `running`.